enable rate limiting of letter sending; max 100 per IP per 24 hours.
After that requests will be rejected. Ideally we would fall back to requiring a CAPTCHA, and set the limit to 10 per IP per 24 hours or there abouts. It currently has to be quite high because we want to avoid accidentally blocking entire companies who route all their traffic through a single proxy.